+++ src/mongoc/mongoc-buffer.c
@@ -163,7 +163,7 @@ _mongoc_buffer_append_from_stream (mongoc_buffer_t *buffer,
       }
       buffer->off = 0;
       if (!SPACE_FOR (buffer, size)) {
-         buffer->datalen = bson_next_power_of_two ((uint32_t)(size + buffer->len + buffer->off));
+         buffer->datalen = bson_next_power_of_two (size + buffer->len + buffer->off);
          buffer->data = buffer->realloc_func (buffer->data, buffer->datalen, NULL);
       }
    }
@@ -231,7 +231,7 @@ _mongoc_buffer_fill (mongoc_buffer_t *buffer,
    buffer->off = 0;
 
    if (!SPACE_FOR (buffer, min_bytes)) {
-      buffer->datalen = bson_next_power_of_two ((uint32_t)(buffer->len + min_bytes));
+      buffer->datalen = bson_next_power_of_two (buffer->len + min_bytes);
       buffer->data = buffer->realloc_func (buffer->data, buffer->datalen,
                                            buffer->realloc_data);
    }
